metrics-streaming-service/requirements.txt
numpy
pandas
pytest
coverage
kafka-python
quantileflow
ddsketch